+++ tests/testthat/test-length.R
@@ -5,6 +5,9 @@
 #' * Result of call will create SAS default length attribute (`width` for each
 #' variable)
 
+# set the testthat edition used by this file
+testthat::local_edition(3)
+
 # xportr_length
 ## Test 1: xportr_length: Accepts valid domain names in metadata object ----
 test_that("length Test 1: Accepts valid domain names in metadata object", {
@@ -157,7 +160,8 @@ test_that("length Test 5: Metacore instance can be used", {
         type = TRUE,
         label = TRUE,
         format = TRUE,
-        order = TRUE
+        order = TRUE,
+        var_names = colnames(adsl)
       )
     )
   ))
@@ -255,4 +259,21 @@ test_that("length Test 11: Works as expected with only one domain in metadata", 
   )
 
   expect_silent(xportr_length(adsl, metadata))
+})
+
+## Test 12: xportr_length: Throws message when metadata variables not present in dataset ----
+test_that("length Test 12: Throws message when metadata variables not present in dataset", {
+  adsl <- data.frame(
+    USUBJID = c(1001, 1002, 1003)
+  )
+
+  metadata <- data.frame(
+    dataset = c("adsl", "adsl"),
+    variable = c("USUBJID", "BRTHDT"),
+    length = c(1, 1)
+  )
+
+  expect_snapshot({
+    xportr_length(adsl, metadata, domain = "adsl", verbose = "message")
+  })
 })