+++ NEWS.md
@@ -3,7 +3,7 @@
 ## New Features
 * Added a check in `xpt_validate()` to verify the class of date, datetime, and
 time variables. (#298)
-* Added reverse check to `xportr_label()` to alert users when metadata contains variables that are not present in the dataset (#284)
+* Added reverse check to `xportr_label()` and `xportr_length()` to alert users when metadata contains variables that are not present in the dataset (#284)
 
 ## Bug Fixes
 * Fixed verbose option bugs in `xportr_format()` and added missing `xportr.order_verbose` option (#318)